infy_tracker/requirements.txt
streamlit>=1.32.0
openai>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.2
plotly>=5.18.0
fpdf2>=2.7.0
